ovsdb-cs: Add coverage for full batch events.
When many update notifications arrive in a short time, it may happen
that we only process the first 50 of them and have to wait until the
next iteration to pick them up.

Because previously, there was no way for an operator to observe how
often this happened, we add a coverage metric for this event.

Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>