jsonrpc: Optimize receive buffer size.
Previously, we reused the jsonrpc byteq with the state from a previous
iteration with potentially less headroom. However, in case we have an
empty byteq, we can reset the pointers to maximize the available headroom
and thus, the number of bytes that we can process in a single iteration.

Co-authored-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
(cherry picked from commit 75844bfc068c82b4ed54b782a2a93f23d1edbf79)