jsonrpc: Optimize receive buffer size.
Previously, we reused the jsonrpc byteq with the state from a previous
iteration with potentially less headroom. However, in case we have an
empty byteq, we can reset the pointers to maximize the available headroom
and thus, the number of bytes that we can process in a single iteration.

Co-authored-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: Martin Morgenstern <martin.morgenstern@cloudandheat.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
(cherry picked from commit 27953b894ca7334b9f0c62a44fd74b66cb9f4018)