python: Disable check for globals with never reassigned names (F824).
flake8 2.7.0 introduced the F824 check [0].

While the check sounds reasonable on the surface, we have code in
the repository where the global variable is used only for access.

My conclusion from evaluating the alternative of adjusting the code
to appease the check is that it would invite future mistakes where
the global variable is unintentionally shadowed in addition to
making the code less clear and explicit.

0: https://github.com/PyCQA/flake8/pull/1974
1: https://github.com/PyCQA/pyflakes/pull/825
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
(cherry picked from commit 7f4ee4395637e906563c71f4112b809fed4f3968)